
\name{read_pgn}
\alias{read_pgn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Read a PGN file}
\description{
  Reads a plaintext file containing chess game information in the Portable Game Notation (PGN) format.
}
\usage{
read_pgn( file )
}
\arguments{
    \item{file}{the path to a plaintext file containing one or more chess game records in PGN format}
}
\details{
    The target file must contain only PGN games, seperated by spaces before the first line and after the last line.
}
%- maybe also 'usage' for other objects documented here.
\references{}
\author{Bret Beheim}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }